Troubleshoot patient update endpoint

+++ HospitalCMS_API/Models/PatientModel.cs
@@ -11,8 +11,8 @@ namespace HospitalCMS_API.Models
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public int Age { get; set; }
-        public double Height { get; set; }
-        public double Weight { get; set; }
+        public decimal Height { get; set; }
+        public decimal Weight { get; set; }
         public string LegalSex { get; set; }
         public string MaritalStatus { get; set; }
         public string Race { get; set; }
